:sparkles: create project
This project is an investigation into using Flutter and Dart for building
Fermilab applications. This commit only contains the initial files that
were generated when making a new project (using Android Studio.) Right now,
this project is only configured to produce a web app. The dialog boxes
promise you can easily add targets later. We'll see if this is true.