:art: tidy the default example
This is still the example when you create a new project. However, I
removed the comments that explain Flutter because this demo should look
like an application we would build and we wouldn't (shouldn't) be
explaining Flutter over and over.  I also renamed the widget to something
more "real-world" and "professional" (i.e.  don't include the word "My" in
them.)